updated auth for create-account and reset-password, removed auth server pages
