+++ src/routes/(auth)/login/+page.svelte
@@ -86,7 +86,7 @@
 </script>
 
 <AuthBase>
-	<h1 class="mb-4 text-4xl font-semibold">Login</h1>
+	<h1 class="mb-8 text-4xl font-semibold">Login</h1>
 	<form method="POST" class="[&>*]:mb-4" onsubmit={handleLogin}>
 		<label class="label">
 			<p>Username or email</p>
