+++ src/routes/(auth)/reset-password/+page.svelte
@@ -2,10 +2,11 @@
 	import EmailField from '$lib/components/EmailField.svelte'
 	import FormError from '$lib/components/FormError.svelte'
 	import { makeSetType } from '$lib/utils.js'
-	import { enhance } from '$app/forms'
 	import type { AuthResult } from '$lib/types'
 	import { onMount } from 'svelte'
 	import AuthBase from '../AuthBase.svelte'
+	import { pb } from '$lib/pocketbase'
+	import { errorMessage } from '$lib/utils'
 
 	let email = $state('')
 	let showEmailValidation = $state(false)
@@ -13,6 +14,13 @@
 	let loading = $state(false)
 	let success = $state(false)
 	let disabled = $derived(loading || showEmailValidation)
+
+	let emailRef: HTMLInputElement | null = $state(null)
+	onMount(() => {
+		if (emailRef) {
+			emailRef.focus()
+		}
+	})
 
 	// focus on the button after correcting email format and tabbing
 	let buttonRef: HTMLButtonElement | null = $state(null)
@@ -23,37 +31,34 @@
 		}
 	})
 
-	const setType = makeSetType<AuthResult>()
+	const handleResetPassword = async (event: Event) => {
+		event.preventDefault()
+		loading = true
+		error = ''
+		success = false
 
-	let emailRef: HTMLInputElement | null = $state(null)
-	onMount(() => {
-		if (emailRef) {
-			emailRef.focus()
+		if (email === '') {
+			error = 'Please enter your email'
+			loading = false
+			return
 		}
-	})
+
+		try {
+			await pb.collection('user').requestPasswordReset(email)
+			success = true
+			error = ''
+		} catch (err) {
+			error = errorMessage(err)
+		} finally {
+			loading = false
+		}
+	}
 </script>
 
 <AuthBase>
-	<h1 class="mb-4 text-4xl font-semibold">Reset Password</h1>
+	<h1 class="mb-8 text-4xl font-semibold">Reset Password</h1>
 	<p class="mb-4">Enter your email and we'll send you a link to reset your password</p>
-	<form
-		method="POST"
-		use:enhance={() => {
-			loading = true
-			error = ''
-			return async ({ result, update }) => {
-				await update()
-				const typedResult = setType(result)
-				if (result.status === 200) {
-					success = true
-				} else {
-					success = false
-					error = typedResult.data.error
-				}
-				loading = false
-			}
-		}}
-	>
+	<form method="POST" onsubmit={handleResetPassword}>
 		<EmailField bind:email bind:showValidation={showEmailValidation} bind:ref={emailRef} />
 		<div class="flex justify-center">
 			<button
